Fix Reconfiguration Window freezing at shutdown
The contents of the window change only if a valid node is selected (i.e.
not the dummy "-" node. Moreover the window sometimes froze when
shutting down due to thread racing conditions where the threadPoller was
trying to acquire the Dispatcher on a closing window. This was fixed bu
adding a timespan timeout to the dispatcher acquisition as well as
moving the login to the "WindowCloasing" event.